+++ build_mkdocs.sh
@@ -1,0 +1,2 @@
+pip install -r requirements-doc.txt
+mkdocs build