+++ src/index.ts
@@ -1,1 +1,6 @@
-export * as Instructor from "./instructor"
+import Instructor, { InstructorClient } from "./instructor"
+
+export { type InstructorClient }
+export * from "./types"
+
+export default Instructor