+++ tests/extract.test.ts
@@ -21,7 +21,7 @@ async function extractUser() {
 
   const user = await client.chat.completions.create({
     messages: [{ role: "user", content: "Jason Liu is 30 years old" }],
-    model: "gpt-3.5-turbo",
+    model: "gpt-4-turbo",
     response_model: { schema: UserSchema, name: "User" },
     seed: 1
   })
@@ -82,7 +82,7 @@ async function extractUserMany() {
 
   const user = await client.chat.completions.create({
     messages: [{ role: "user", content: "Jason is 30 years old, Sarah is 12" }],
-    model: "gpt-3.5-turbo",
+    model: "gpt-4-turbo",
     response_model: { schema: UsersSchema, name: "Users" },
     max_retries: 3,
     seed: 1