+++ tests/inference.test.ts
@@ -7,9 +7,10 @@
 // 6. response_model, no stream, max_retries
 
 import Instructor from "@/instructor"
+import { type CompletionMeta } from "@/types"
 import { describe, expect, test } from "bun:test"
 import OpenAI from "openai"
-import { Stream } from "openai/streaming.mjs"
+import { Stream } from "openai/streaming"
 import { type } from "ts-inference-check"
 import { z } from "zod"
 
@@ -32,7 +33,7 @@ describe("Inference Checking", () => {
   test("no response_model, no stream", async () => {
     const user = await client.chat.completions.create({
       messages: [{ role: "user", content: "Jason Liu is 30 years old" }],
-      model: "gpt-3.5-turbo",
+      model: "gpt-4-turbo",
       seed: 1,
       stream: false
     })
@@ -43,30 +44,34 @@ describe("Inference Checking", () => {
   test("no response_model, stream", async () => {
     const userStream = await client.chat.completions.create({
       messages: [{ role: "user", content: "Jason Liu is 30 years old" }],
-      model: "gpt-3.5-turbo",
+      model: "gpt-4-turbo",
       seed: 1,
       stream: true
     })
 
-    expect(type(userStream).strictly.is<Stream<OpenAI.Chat.ChatCompletionChunk>>(true)).toBe(true)
+    expect(
+      type(userStream).strictly.is<Stream<OpenAI.Chat.Completions.ChatCompletionChunk>>(true)
+    ).toBe(true)
   })
 
   test("response_model, no stream", async () => {
     const user = await client.chat.completions.create({
       messages: [{ role: "user", content: "Jason Liu is 30 years old" }],
-      model: "gpt-3.5-turbo",
+      model: "gpt-4-turbo",
       response_model: { schema: UserSchema, name: "User" },
       seed: 1,
       stream: false
     })
 
-    expect(type(user).strictly.is<z.infer<typeof UserSchema>>(true)).toBe(true)
+    expect(
+      type(user).strictly.is<z.infer<typeof UserSchema> & { _meta?: CompletionMeta }>(true)
+    ).toBe(true)
   })
 
   test("response_model, stream", async () => {
     const userStream = await client.chat.completions.create({
       messages: [{ role: "user", content: "Jason Liu is 30 years old" }],
-      model: "gpt-3.5-turbo",
+      model: "gpt-4-turbo",
       response_model: { schema: UserSchema, name: "User" },
       seed: 1,
       stream: true
@@ -78,7 +83,7 @@ describe("Inference Checking", () => {
           Partial<{
             name: string
             age: number
-          }>,
+          }> & { _meta?: CompletionMeta },
           void,
           unknown
         >
@@ -89,7 +94,7 @@ describe("Inference Checking", () => {
   test("response_model, stream, max_retries", async () => {
     const userStream = await client.chat.completions.create({
       messages: [{ role: "user", content: "Jason Liu is 30 years old" }],
-      model: "gpt-3.5-turbo",
+      model: "gpt-4-turbo",
       response_model: { schema: UserSchema, name: "User" },
       seed: 1,
       stream: true,
@@ -102,7 +107,7 @@ describe("Inference Checking", () => {
           Partial<{
             name: string
             age: number
-          }>,
+          }> & { _meta?: CompletionMeta },
           void,
           unknown
         >
@@ -113,12 +118,14 @@ describe("Inference Checking", () => {
   test("response_model, no stream, max_retries", async () => {
     const user = await client.chat.completions.create({
       messages: [{ role: "user", content: "Jason Liu is 30 years old" }],
-      model: "gpt-3.5-turbo",
+      model: "gpt-4-turbo",
       response_model: { schema: UserSchema, name: "User" },
       seed: 1,
       max_retries: 3
     })
 
-    expect(type(user).strictly.is<z.infer<typeof UserSchema>>(true)).toBe(true)
+    expect(
+      type(user).strictly.is<z.infer<typeof UserSchema> & { _meta?: CompletionMeta }>(true)
+    ).toBe(true)
   })
 })