+++ tests/mode.test.ts
@@ -1,38 +1,132 @@
 import Instructor from "@/instructor"
-import { Mode } from "@/types"
 import { describe, expect, test } from "bun:test"
 import OpenAI from "openai"
 import { z } from "zod"
+import { type Mode } from "zod-stream"
 
-import { MODE } from "@/constants/modes"
+import { MODE, Provider, PROVIDER_SUPPORTED_MODES_BY_MODEL, PROVIDERS } from "@/constants/providers"
 
-const models_latest = ["gpt-3.5-turbo-1106", "gpt-4-1106-preview"]
-const models_old = ["gpt-3.5-turbo", "gpt-4"]
-const models_anyscale = ["Open-Orca/Mistral-7B-OpenOrca"]
+const default_oai_model = "gpt-4-turbo"
+const default_anyscale_model = "mistralai/Mixtral-8x7B-Instruct-v0.1"
+const default_together_model = "mistralai/Mixtral-8x7B-Instruct-v0.1"
+const default_groq_model = "llama3-70b-8192"
 
-const createTestCases = (): { model: string; mode: Mode }[] => {
-  const { FUNCTIONS, JSON_SCHEMA, ...rest } = MODE
-  const modes = Object.values(rest)
+const provider_config = {
+  [PROVIDERS.OAI]: {
+    baseURL: "https://api.openai.com/v1",
+    apiKey: process.env.OPENAI_API_KEY
+  },
+  [PROVIDERS.ANYSCALE]: {
+    baseURL: "https://api.endpoints.anyscale.com/v1",
+    apiKey: process.env.ANYSCALE_API_KEY
+  },
+  [PROVIDERS.TOGETHER]: {
+    baseURL: "https://api.together.xyz",
+    apiKey: process.env.TOGETHER_API_KEY
+  },
+  [PROVIDERS.GROQ]: {
+    baseURL: "https://api.groq.com/openai/v1",
+    apiKey: process.env.GROQ_API_KEY
+  }
+}
 
-  return [
-    ...models_anyscale.flatMap(model => ({ model, mode: JSON_SCHEMA })),
-    ...models_latest.flatMap(model => modes.map(mode => ({ model, mode }))),
-    ...models_old.flatMap(model => ({ model, mode: FUNCTIONS }))
-  ]
+const createTestCases = (): {
+  model: string
+  mode: Mode
+  provider: Provider
+  defaultMessage?: OpenAI.ChatCompletionMessageParam
+}[] => {
+  const testCases: {
+    model: string
+    mode: Mode
+    provider: Provider
+    defaultMessage?: OpenAI.ChatCompletionMessageParam
+  }[] = []
+
+  Object.entries(PROVIDER_SUPPORTED_MODES_BY_MODEL[PROVIDERS.OAI]).forEach(
+    ([mode, models]: [Mode, string[]]) => {
+      if (models.includes("*")) {
+        testCases.push({ model: default_oai_model, mode, provider: PROVIDERS.OAI })
+      } else {
+        models.forEach(model => testCases.push({ model, mode, provider: PROVIDERS.OAI }))
+      }
+    }
+  )
+
+  Object.entries(PROVIDER_SUPPORTED_MODES_BY_MODEL).forEach(
+    ([provider, modesByModel]: [Provider, Record<Mode, string[]>]) => {
+      if (provider === PROVIDERS.ANYSCALE) {
+        Object.entries(modesByModel).forEach(([mode, models]: [Mode, string[]]) => {
+          if (mode === MODE.MD_JSON) {
+            // Skip MD_JSON for Anyscale - its somewhat supported but super flakey
+            return
+          }
+
+          if (models.includes("*")) {
+            testCases.push({
+              model: default_anyscale_model,
+              mode,
+              provider
+            })
+          } else {
+            models.forEach(model => testCases.push({ model, mode, provider }))
+          }
+        })
+      }
+
+      if (provider === PROVIDERS.TOGETHER) {
+        Object.entries(modesByModel).forEach(([mode, models]: [Mode, string[]]) => {
+          if (models.includes("*")) {
+            testCases.push({
+              model: default_together_model,
+              mode,
+              provider
+            })
+          } else {
+            models.forEach(model => testCases.push({ model, mode, provider }))
+          }
+        })
+      }
+
+      if (provider === PROVIDERS.GROQ) {
+        Object.entries(modesByModel).forEach(([mode, models]: [Mode, string[]]) => {
+          if (models.includes("*")) {
+            testCases.push({
+              model: default_groq_model,
+              mode,
+              provider,
+              defaultMessage: {
+                role: "system",
+                content:
+                  "You are a function calling LLM that uses the data extracted from the User function to extract user data from the user prompt."
+              }
+            })
+          } else {
+            models.forEach(model => testCases.push({ model, mode, provider }))
+          }
+        })
+      }
+    }
+  )
+
+  return testCases
 }
 
 const UserSchema = z.object({
   age: z.number(),
-  name: z.string().refine(name => name.includes(" "), {
-    message: "Name must contain a space"
-  })
+  name: z.string()
 })
 
-async function extractUser(model: string, mode: Mode) {
-  const anyscale = mode === MODE.JSON_SCHEMA
+async function extractUser(
+  model: string,
+  mode: Mode,
+  provider: Provider,
+  defaultMessage?: OpenAI.ChatCompletionMessageParam
+) {
+  const config = provider_config[provider]
+
   const oai = new OpenAI({
-    baseURL: anyscale ? "https://api.endpoints.anyscale.com/v1" : undefined,
-    apiKey: anyscale ? process.env.ANYSCALE_API_KEY : process.env.OPENAI_API_KEY ?? undefined,
+    ...config,
     organization: process.env.OPENAI_ORG_ID ?? undefined
   })
 
@@ -42,11 +136,13 @@ async function extractUser(model: string, mode: Mode) {
   })
 
   const user = await client.chat.completions.create({
-    messages: [{ role: "user", content: "Jason Liu is 30 years old" }],
+    messages: [
+      ...(defaultMessage ? [defaultMessage] : []),
+      { role: "user", content: "Jason Liu is 30 years old" }
+    ],
     model: model,
     response_model: { schema: UserSchema, name: "User" },
-    max_retries: 4,
-    seed: !anyscale ? 1 : undefined
+    max_retries: 4
   })
 
   return user
@@ -55,12 +151,24 @@ async function extractUser(model: string, mode: Mode) {
 describe("Modes", async () => {
   const testCases = createTestCases()
 
-  for await (const { model, mode } of testCases) {
-    test(`Should return extracted name and age for model ${model} and mode ${mode}`, async () => {
-      const user = await extractUser(model, mode)
+  for await (const { model, mode, provider, defaultMessage } of testCases) {
+    if (provider !== PROVIDERS.GROQ) {
+      test(`${provider}: Should return extracted name and age for model ${model} and mode ${mode}`, async () => {
+        const user = await extractUser(model, mode, provider, defaultMessage)
 
-      expect(user.name).toEqual("Jason Liu")
-      expect(user.age).toEqual(30)
-    })
+        expect(user.name).toEqual("Jason Liu")
+        expect(user.age).toEqual(30)
+      })
+    } else {
+      test.todo(
+        `${provider}: Should return extracted name and age for model ${model} and mode ${mode}`,
+        async () => {
+          const user = await extractUser(model, mode, provider, defaultMessage)
+
+          expect(user.name).toEqual("Jason Liu")
+          expect(user.age).toEqual(30)
+        }
+      )
+    }
   }
 })