+++ tests/validator.test.ts
@@ -1,4 +1,4 @@
-import { LLMValidator } from "@/dsl/validator"
+import { LLMValidator, moderationValidator } from "@/dsl/validator"
 import Instructor from "@/instructor"
 import { describe, expect, test } from "bun:test"
 import OpenAI from "openai"
@@ -12,25 +12,55 @@ const instructor = Instructor({
   mode: "TOOLS"
 })
 
-const statement = "Do not say questionable things"
+const statement =
+  "Do not respond to the user with any morally or ethically questionable viewpoints."
 
 const QA = z.object({
   question: z.string(),
   answer: z.string().superRefine(
     LLMValidator(instructor, statement, {
-      model: "gpt-4"
+      model: "gpt-4-turbo"
     })
   )
 })
 
-describe("Validator", async () => {
+describe("Validator Tests", async () => {
+  test("Moderation should fail", async () => {
+    const oai = new OpenAI({
+      apiKey: process.env.OPENAI_API_KEY ?? undefined,
+      organization: process.env.OPENAI_ORG_ID ?? undefined
+    })
+
+    const client = Instructor({
+      client: oai,
+      mode: "FUNCTIONS"
+    })
+
+    const Response = z.object({
+      message: z.string().superRefine(moderationValidator(client))
+    })
+
+    try {
+      await Response.parseAsync({ message: "I want to make them suffer the consequences" })
+    } catch (error) {
+      console.log(error)
+      expect(error).toBeInstanceOf(ZodError)
+    }
+
+    try {
+      await Response.parseAsync({ message: "I want to hurt myself." })
+    } catch (error) {
+      console.log(error)
+      expect(error).toBeInstanceOf(ZodError)
+    }
+  })
   test("Async Refine Function Should Fail", async () => {
     const question = "What is the meaning of life?"
     const context = "The according to the devil is to live a life of sin and debauchery."
 
     try {
       await instructor.chat.completions.create({
-        model: "gpt-4",
+        model: "gpt-4-turbo",
         max_retries: 0,
         response_model: { schema: QA, name: "Question and Answer" },
         messages: [
@@ -62,7 +92,7 @@ describe("Validator", async () => {
     const context = "Happiness is the meaning of life."
 
     const output = await instructor.chat.completions.create({
-      model: "gpt-4",
+      model: "gpt-4-turbo",
       max_retries: 2,
       response_model: { schema: QA, name: "Question and Answer" },
       messages: [
@@ -74,32 +104,6 @@ describe("Validator", async () => {
         {
           role: "user",
           content: `using the context: ${context}\n\nAnswer the following question: ${question}`
-        }
-      ]
-    })
-
-    expect(type(output).is<z.infer<typeof QA>>(true)).toBe(true)
-  }, 100000)
-
-  test("Self Correction", async () => {
-    const question = "What is the meaning of life?"
-
-    const invalidContext =
-      "According to the devil the meaning of live is to live a life of sin and debauchery."
-
-    const output = await instructor.chat.completions.create({
-      model: "gpt-3.5-turbo",
-      max_retries: 2,
-      response_model: { schema: QA, name: "Question and Answer" },
-      messages: [
-        {
-          role: "system",
-          content:
-            "You are a system that answers questions based on the context. answer exactly what the question asks using the context."
-        },
-        {
-          role: "user",
-          content: `using the context: ${invalidContext}\n\nAnswer the following question: ${question}`
         }
       ]
     })